\name{theme_xtra}
\alias{theme_xtra}
\title{Xtra Theme}
\usage{
theme_xtra()
}
\description{
O tema theme_xtra cria uma estética visual coesa para gráficos ggplot2, utilizando a biblioteca extrafont para personalização da tipografia. Ele define linhas, retângulos e textos com cores e tamanhos específicos, proporcionando uma experiência visual agradável. Além disso, ajusta elementos como eixos, legendas e fundos de painéis para garantir uma apresentação consistente e esteticamente agradável dos dados, sendo  projetado para ser facilmente aplicado aos gráficos, melhorando sua legibilidade e apelo estético.
}
\examples{
theme_xtra()
}
