\name{theme_academic}
\alias{theme_academic}
\title{Academic Theme}
\usage{
theme_academic()
}
\description{
O tema acadêmico simplificado foi projetado para proporcionar uma experiência visual clara e direta ao apresentar gráficos no meio acadêmico. Com bordas sutis apenas nos eixos X e Y, elimina distrações desnecessárias, concentrando-se na comunicação eficaz de dados. Detalhes cuidadosamente ajustados, como cores suaves e espaçamento equilibrado, garantem uma apresentação limpa, tornando a interpretação e análise dos gráficos uma tarefa intuitiva.
}
\examples{
theme_academic()
}
